requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
df2gspread==1.0.4
ecdsa==0.19.0
enum-compat==0.0.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gspread==5.12.3
httplib2==0.22.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==5.13.0
inflection==0.5.1
intuit-oauth==1.2.5
mysql-connector-python==8.2.0
numpy==1.26.2
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
proto-plus==1.26.1
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyodbc==5.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-jose==3.3.0
python-quickbooks==0.9.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
tabulate==0.9.0
tqdm==4.66.1
tzdata==2023.3
uritemplate==3.0.1
urllib3==2.1.0
zipp==3.17.0
